Cuda merge with master

Former-commit-id: 766b5441f9d82eab33828e50a5624f7b10364c30 [formerly 2740f642dfef8c1d60dea6d914aea5106f7b2314] [formerly 85b25b9e1cec0663b0375e3e15ca8ca8fe5819ae [formerly acfa07be517ba4e2b66ec3c3b80e776b3f861bc5]]
Former-commit-id: 1fa708bff1e69d1e24533966c22f330374fc518b [formerly 7d5575c4233171cf99a3a7fa5ce8da0c40076611]
Former-commit-id: 059265710691126b7f888b80f464ead94b0258cc

+++ src/eval/evaluation.h
@@ -26,14 +26,14 @@ namespace FT{
          * @class Evaluation
          * @brief evaluation mixin class for Feat
          */
-        typedef double (*funcPointer)(const VectorXd&, const shared_ptr<CLabels>&, VectorXd&,
+        typedef float (*funcPointer)(const VectorXf&, const shared_ptr<CLabels>&, VectorXf&,
                                       const vector<float>&);
         
         class Evaluation 
         {
             public:
             
-                double (* score)(const VectorXd&, const shared_ptr<CLabels>&, VectorXd&, 
+                float (* score)(const VectorXf&, const shared_ptr<CLabels>&, VectorXf&, 
                                  const vector<float>&);    // pointer to scoring function
                                  
                 std::map<string, funcPointer> score_hash;
@@ -47,14 +47,14 @@ namespace FT{
                 /// fitness of population.
                 void fitness(vector<Individual>& individuals,
                              const Data& d, 
-                             MatrixXd& F, 
+                             MatrixXf& F, 
                              const Parameters& params, 
                              bool offspring = false,
                              bool validation = false);
               
                 /// assign fitness to an individual and to F.  
-                void assign_fit(Individual& ind, MatrixXd& F, const shared_ptr<CLabels>& yhat, 
-                                const VectorXd& y, const Parameters& params,bool val=false);       
+                void assign_fit(Individual& ind, MatrixXf& F, const shared_ptr<CLabels>& yhat, 
+                                const VectorXf& y, const Parameters& params,bool val=false);       
         };
     }
 }
