+++ src/sel/selection_operator.cc
@@ -11,12 +11,12 @@ namespace FT{
 
         SelectionOperator::~SelectionOperator(){}
         
-        vector<size_t> SelectionOperator::select(Population& pop, const MatrixXd& F, const Parameters& p) 
+        vector<size_t> SelectionOperator::select(Population& pop, const MatrixXf& F, const Parameters& p) 
         {   
             HANDLE_ERROR_THROW("Undefined select() operation");
         }
         
-        vector<size_t> SelectionOperator::survive(Population& pop, const MatrixXd& F, const Parameters& p)
+        vector<size_t> SelectionOperator::survive(Population& pop, const MatrixXf& F, const Parameters& p)
         {
             HANDLE_ERROR_THROW("Undefined select() operation");
         }
