+++ tests/populationTests.cc
@@ -6,8 +6,8 @@ TEST(Population, PopulationTests)
     feat.set_n_threads(1);
 	feat.set_random_state(666);
 	
-    MatrixXd X(4,2); 
-    MatrixXd X_v(3,2); 
+    MatrixXf X(4,2); 
+    MatrixXf X_v(3,2); 
     X << 0,1,  
          0.47942554,0.87758256,  
          0.84147098,  0.54030231,
@@ -19,13 +19,13 @@ TEST(Population, PopulationTests)
     X.transposeInPlace();
     X_v.transposeInPlace();
 
-    VectorXd y(4); 
-    VectorXd y_v(3); 
+    VectorXf y(4); 
+    VectorXf y_v(3); 
     // y = 2*x1 + 3.x2
     y << 3.0,  3.59159876,  3.30384889,  2.20720158;
     y_v << 0.57015434, -1.20648656, -2.68773747;
     
-    std::map<string, std::pair<vector<ArrayXd>, vector<ArrayXd> > > z;
+    std::map<string, std::pair<vector<ArrayXf>, vector<ArrayXf> > > z;
 
     feat.params.init();
